feat(Turnips): add Turnips chart part 6
* Extracted Chart Grid View and renamed Chart Legend View properly
